# airbyte_cdk/manifest_server/openapi.yaml
# This file is auto-generated. Do not edit manually.
# To regenerate, run: manifest-server generate-openapi

openapi: 3.1.0
info:
  title: Manifest Server Service
  description: A service for running low-code Airbyte connectors
  contact:
    name: Airbyte
    url: https://airbyte.com/
  version: 0.1.0
paths:
  /health/:
    get:
      tags:
      - health
      summary: Health
      operationId: health_health__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /capabilities/:
    get:
      tags:
      - capabilities
      summary: Get Capabilities
      description: "Get the capabilities available for the manifest server service.\n\
        \nReturns:\n    Dict containing the service capabilities including custom\
        \ code execution support."
      operationId: getCapabilities
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Getcapabilities
  /v1/manifest/test_read:
    post:
      tags:
      - manifest
      summary: Test Read
      description: Test reading from a specific stream in the manifest.
      operationId: testRead
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamTestReadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/manifest/check:
    post:
      tags:
      - manifest
      summary: Check
      description: Check configuration against a manifest
      operationId: check
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/manifest/discover:
    post:
      tags:
      - manifest
      summary: Discover
      description: Discover streams from a manifest
      operationId: discover
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoverRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoverResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/manifest/resolve:
    post:
      tags:
      - manifest
      summary: Resolve
      description: Resolve a manifest to its final configuration.
      operationId: resolve
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManifestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/manifest/full_resolve:
    post:
      tags:
      - manifest
      summary: Full Resolve
      description: 'Fully resolve a manifest including dynamic streams.


        Generates dynamic streams up to the specified limit and includes

        them in the resolved manifest.'
      operationId: fullResolve
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullResolveRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManifestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
components:
  schemas:
    AirbyteCatalog:
      properties:
        streams:
          items:
            $ref: '#/components/schemas/AirbyteStream'
          type: array
          title: Streams
      type: object
      required:
      - streams
      title: AirbyteCatalog
    AirbyteStream:
      properties:
        name:
          type: string
          title: Name
        json_schema:
          additionalProperties: true
          type: object
          title: Json Schema
        supported_sync_modes:
          items:
            $ref: '#/components/schemas/SyncMode'
          type: array
          title: Supported Sync Modes
        source_defined_cursor:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Source Defined Cursor
        default_cursor_field:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Default Cursor Field
        source_defined_primary_key:
          anyOf:
          - items:
              items:
                type: string
              type: array
            type: array
          - type: 'null'
          title: Source Defined Primary Key
        namespace:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace
        is_resumable:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Resumable
        is_file_based:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is File Based
      type: object
      required:
      - name
      - json_schema
      - supported_sync_modes
      title: AirbyteStream
    AuxiliaryRequest:
      properties:
        title:
          type: string
          title: Title
        type:
          type: string
          title: Type
        description:
          type: string
          title: Description
        request:
          $ref: '#/components/schemas/HttpRequest'
        response:
          $ref: '#/components/schemas/HttpResponse'
      type: object
      required:
      - title
      - type
      - description
      - request
      - response
      title: AuxiliaryRequest
      description: Auxiliary HTTP request made during stream processing.
    CheckRequest:
      properties:
        manifest:
          $ref: '#/components/schemas/Manifest'
        config:
          $ref: '#/components/schemas/ConnectorConfig'
      type: object
      required:
      - manifest
      - config
      title: CheckRequest
      description: Request to check a manifest.
    CheckResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
      type: object
      required:
      - success
      title: CheckResponse
      description: Response to check a manifest.
    ConnectorConfig:
      properties: {}
      additionalProperties: true
      type: object
      title: ConnectorConfig
      description: Base connector configuration model. Allows client generation to
        replace with proper JsonNode types.
    DiscoverRequest:
      properties:
        manifest:
          $ref: '#/components/schemas/Manifest'
        config:
          $ref: '#/components/schemas/ConnectorConfig'
      type: object
      required:
      - manifest
      - config
      title: DiscoverRequest
      description: Request to discover a manifest.
    DiscoverResponse:
      properties:
        catalog:
          $ref: '#/components/schemas/AirbyteCatalog'
      type: object
      required:
      - catalog
      title: DiscoverResponse
      description: Response to discover a manifest.
    FullResolveRequest:
      properties:
        manifest:
          $ref: '#/components/schemas/Manifest'
        config:
          $ref: '#/components/schemas/ConnectorConfig'
        stream_limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Stream Limit
          default: 100
      type: object
      required:
      - manifest
      - config
      title: FullResolveRequest
      description: Request to fully resolve a manifest.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HttpRequest:
      properties:
        url:
          type: string
          title: Url
        headers:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Headers
        http_method:
          type: string
          title: Http Method
        body:
          anyOf:
          - type: string
          - type: 'null'
          title: Body
      type: object
      required:
      - url
      - headers
      - http_method
      title: HttpRequest
      description: HTTP request details.
    HttpResponse:
      properties:
        status:
          type: integer
          title: Status
        body:
          anyOf:
          - type: string
          - type: 'null'
          title: Body
        headers:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Headers
      type: object
      required:
      - status
      title: HttpResponse
      description: HTTP response details.
    LogMessage:
      properties:
        message:
          type: string
          title: Message
        level:
          type: string
          title: Level
        internal_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Internal Message
        stacktrace:
          anyOf:
          - type: string
          - type: 'null'
          title: Stacktrace
      type: object
      required:
      - message
      - level
      title: LogMessage
      description: Log message from stream processing.
    Manifest:
      properties: {}
      additionalProperties: true
      type: object
      title: Manifest
      description: Base manifest model. Allows client generation to replace with proper
        JsonNode types.
    ManifestResponse:
      properties:
        manifest:
          $ref: '#/components/schemas/Manifest'
      type: object
      required:
      - manifest
      title: ManifestResponse
      description: Response containing a manifest.
    ResolveRequest:
      properties:
        manifest:
          $ref: '#/components/schemas/Manifest'
      type: object
      required:
      - manifest
      title: ResolveRequest
      description: Request to resolve a manifest.
    StreamRead:
      properties:
        logs:
          items:
            $ref: '#/components/schemas/LogMessage'
          type: array
          title: Logs
        slices:
          items:
            $ref: '#/components/schemas/StreamReadSlices'
          type: array
          title: Slices
        test_read_limit_reached:
          type: boolean
          title: Test Read Limit Reached
        auxiliary_requests:
          items:
            $ref: '#/components/schemas/AuxiliaryRequest'
          type: array
          title: Auxiliary Requests
        inferred_schema:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Inferred Schema
        inferred_datetime_formats:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Inferred Datetime Formats
        latest_config_update:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Latest Config Update
      type: object
      required:
      - logs
      - slices
      - test_read_limit_reached
      - auxiliary_requests
      - inferred_schema
      - inferred_datetime_formats
      - latest_config_update
      title: StreamRead
      description: Complete stream read response with properly typed fields.
    StreamReadPages:
      properties:
        records:
          items: {}
          type: array
          title: Records
        request:
          anyOf:
          - $ref: '#/components/schemas/HttpRequest'
          - type: 'null'
        response:
          anyOf:
          - $ref: '#/components/schemas/HttpResponse'
          - type: 'null'
      type: object
      required:
      - records
      title: StreamReadPages
      description: Pages of data read from a stream slice.
    StreamReadSlices:
      properties:
        pages:
          items:
            $ref: '#/components/schemas/StreamReadPages'
          type: array
          title: Pages
        slice_descriptor:
          anyOf:
          - type: string
          - type: 'null'
          title: Slice Descriptor
        state:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: State
        auxiliary_requests:
          anyOf:
          - items:
              $ref: '#/components/schemas/AuxiliaryRequest'
            type: array
          - type: 'null'
          title: Auxiliary Requests
      type: object
      required:
      - pages
      - slice_descriptor
      title: StreamReadSlices
      description: Slices of data read from a stream.
    StreamTestReadRequest:
      properties:
        manifest:
          $ref: '#/components/schemas/Manifest'
        config:
          $ref: '#/components/schemas/ConnectorConfig'
        stream_name:
          type: string
          title: Stream Name
        state:
          items: {}
          type: array
          title: State
          default: []
        custom_components_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Custom Components Code
        record_limit:
          type: integer
          maximum: 5000.0
          minimum: 1.0
          title: Record Limit
          default: 100
        page_limit:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Page Limit
          default: 5
        slice_limit:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Slice Limit
          default: 5
      type: object
      required:
      - manifest
      - config
      - stream_name
      title: StreamTestReadRequest
      description: Request to test read from a specific stream.
    SyncMode:
      type: string
      enum:
      - full_refresh
      - incremental
      title: SyncMode
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer